Correcting biased spherical sampling in ALT files
Previous algorithm resulted in incorrectly distributed points in the surface of the sphere (for a given radius r), with higher density near the poles. The new function *mutate* gets homogeneously distributed points in the surface of a sphere with radius r. Still, as we are not normalizing with respect to the radius, the density will be higher near the center (i.e. parental genotype), which is a desired feature for us.